+++ src/Shrun/Logging/Queue.hs
@@ -1,4 +1,3 @@
-{-# LANGUAGE TemplateHaskell #-}
 {-# LANGUAGE UndecidableInstances #-}
 
 -- | Provides the 'LogTextQueue' type and associated functions. This is intended
@@ -8,7 +7,7 @@
 module Shrun.Logging.Queue
   ( -- * LogText
     LogText (MkLogText),
-    _LogText,
+    _MkLogText,
     formatFileLog,
 
     -- * Queue
@@ -20,71 +19,44 @@ module Shrun.Logging.Queue
   )
 where
 
-import Data.Text qualified as T
+import Shrun.Configuration.Env.Types (HasLogging (..))
 import Shrun.Effects.Mutable (Mutable (..))
 import Shrun.Effects.Timing (Timing (..))
-import Shrun.Logging.Types (Log (..))
+import Shrun.Logging.Formatting (stripChars')
+import Shrun.Logging.Types
+  ( Log (..),
+    LogText (..),
+    LogTextQueue (..),
+    _MkLogText,
+    _MkLogTextQueue,
+  )
 import Shrun.Logging.Types qualified as Log
 import Shrun.Prelude
-
--- $setup
--- >>> import Shrun.Logging.Types (LogDest (..), LogLevel (..), LogMode (..))
--- >>> import Data.Text qualified as T
--- >>> :{
---  hardcodeTimestamp :: LogText -> Text
---  hardcodeTimestamp (MkLogText txt) = ts <> dropTimestamp txt
---    where
---      dropTimestamp = T.dropWhile (/= ']')
---      ts = "[2022-02-23 20:58:04.231933782 UTC"
--- :}
-
--- | 'LogText' is a textual representation of a given 'Log'. No coloring
--- is included, but we include the prefix (e.g. Warn) along with a timestamp.
---
--- @since 0.1
-newtype LogText = UnsafeLogText
-  { -- | @since 0.1
-    unLogText :: Text
-  }
-  deriving stock
-    ( -- | @since 0.1
-      Eq,
-      -- | @since 0.1
-      Show
-    )
 
 -- | @since 0.5
 _LogText :: Getter LogText Text
 _LogText = to (\(MkLogText t) -> t)
 
--- | @since 0.1
-pattern MkLogText :: Text -> LogText
-pattern MkLogText t <- UnsafeLogText t
-
-{-# COMPLETE MkLogText #-}
-
 -- | Formats a 'Log' into a 'LogText'. Applies prefix and timestamp.
 --
 -- ==== __Examples__
 --
--- >>> :{
---   let log = MkLog Nothing "Running time: 2 seconds" Info Set LogBoth
---    -- timestamp hardcoded for testing
---    in fmap hardcodeTimestamp (formatFileLog log)
--- :}
+-- @
+-- formatFileLog $ MkLog Nothing "Running time: 2 seconds" Info Set LogBoth
 -- "[2022-02-23 20:58:04.231933782 UTC] [Info] Running time: 2 seconds\n"
+-- @
 --
--- >>> :{
---   let log = MkLog (Just "cmd") "cmd: command not found" Error Set LogBoth
---    in fmap hardcodeTimestamp (formatFileLog log)
--- :}
+-- @
+-- formatFileLog $ MkLog (Just "cmd") "cmd: command not found" Error Set LogBoth
 -- "[2022-02-23 20:58:04.231933782 UTC] [Error] [cmd] cmd: command not found\n"
+-- @
 --
 -- @since 0.1
-formatFileLog :: Timing m => Log -> m LogText
+formatFileLog :: (HasLogging env, MonadReader env m, Timing m) => Log -> m LogText
 formatFileLog log = do
   currTime <- getSystemTime
-  let msg' = T.strip $ log ^. #msg
+  stripControl <- asks getFileLogStripControl
+  let msg' = stripChars' (log ^. #msg) stripControl
       formatted = case log ^. #cmd of
         Nothing -> prefix <> msg'
         Just com -> prefix <> "[" <> (com ^. #command) <> "] " <> msg'
@@ -94,26 +66,18 @@ formatFileLog log = do
     prefix = Log.logToPrefix log
 {-# INLINEABLE formatFileLog #-}
 
--- | Newtype wrapper over a 'TBQueue'.
---
--- @since 0.1
-newtype LogTextQueue = MkLogTextQueue
-  { -- | @since 0.1
-    getLogTextQueue :: TBQueue LogText
-  }
-
--- | @since 0.5
-makePrisms ''LogTextQueue
-
--- | @since 0.1
-instance Show LogTextQueue where
-  show _ = "<MkLogTextQueue>"
-  {-# INLINEABLE show #-}
-
 -- | Atomically writes to the queue.
 --
 -- @since 0.1
-writeQueue :: (Mutable m, Timing m) => LogTextQueue -> Log -> m ()
+writeQueue ::
+  ( HasLogging env,
+    MonadReader env m,
+    Mutable m,
+    Timing m
+  ) =>
+  LogTextQueue ->
+  Log ->
+  m ()
 writeQueue queue = writeq <=< formatFileLog
   where
     writeq = liftSTM . writeTBQueue (queue ^. _MkLogTextQueue)