+++ test/unit/Unit/Props/Shrun/Logging/Queue.hs
@@ -15,6 +15,7 @@ import Data.String (IsString)
 import Data.Text qualified as T
 import Hedgehog qualified as H
 import Refined qualified as R
+import Shrun.Configuration.Env.Types (CmdDisplay (..), CmdLogging (..), HasLogging (..), StripControl (..))
 import Shrun.Data.Command (Command (..))
 import Shrun.Effects.Timing (Timing (..))
 import Shrun.Logging.Queue (LogText (..))
@@ -40,6 +41,22 @@ sysTime = "2022-02-20 23:47:39.90228065 UTC"
 sysTimeNE :: Refined R.NonEmpty Text
 sysTimeNE = $$(R.refineTH "[2022-02-20 23:47:39.90228065 UTC]")
 
+newtype MockEnv = MkMockEnv ()
+
+instance HasLogging MockEnv where
+  getCmdDisplay = const ShowKey
+  getCmdLineTrunc = const Nothing
+  getCmdLogging = const Disabled
+  getCmdNameTrunc = const Nothing
+  getFileLogging = const Nothing
+
+  -- This is what we need to run the tests. It's set to
+  -- None since our generators can generate null bytes,
+  -- thus properties would fail unless we prevent that.
+  getFileLogStripControl = const StripControlNone
+  getDisableLogging = const False
+  getStripControl = const StripControlSmart
+
 -- Monad with mock implementation for 'Timing'.
 newtype MockTime a = MkMockTime
   { runMockTime :: a
@@ -50,6 +67,10 @@ newtype MockTime a = MkMockTime
 instance Timing MockTime where
   getSystemTime = pure $ TR.read sysTime
   getTimeSpec = pure $ fromIntegral @Int64 0
+
+instance MonadReader MockEnv MockTime where
+  ask = pure $ MkMockEnv ()
+  local _ = id
 
 makeFieldLabelsNoPrefix ''MockTime
 
@@ -78,7 +99,7 @@ timestampProps = T.askOption $ \(MkMaxRuns limit) ->
     H.withTests limit $
       H.property $ do
         log <- H.forAll LGens.genLog
-        let MkLogText result = Queue.formatFileLog @MockTime log ^. #runMockTime
+        let MkLogText result = Queue.formatFileLog @_ @MockTime log ^. #runMockTime
             (res, _) = Utils.breakStripPoint sysTimeNE result
         "" === res
 
@@ -88,7 +109,7 @@ messageProps = T.askOption $ \(MkMaxRuns limit) ->
     H.withTests limit $
       H.property $ do
         log@MkLog {msg} <- H.forAll LGens.genLog
-        let MkLogText result = Queue.formatFileLog @MockTime log ^. #runMockTime
+        let MkLogText result = Queue.formatFileLog @_ @MockTime log ^. #runMockTime
         H.annotate $ T.unpack result
         H.assert $ T.isInfixOf (T.strip msg) result
 
@@ -98,7 +119,7 @@ prefixProps = T.askOption $ \(MkMaxRuns limit) ->
     H.withTests limit $
       H.property $ do
         log@MkLog {lvl} <- H.forAll LGens.genLog
-        let MkLogText result = Queue.formatFileLog @MockTime log ^. #runMockTime
+        let MkLogText result = Queue.formatFileLog @_ @MockTime log ^. #runMockTime
         case lvl of
           -- level is None: no prefix
           None -> foldr (noMatch result) (pure ()) nonEmptyPrefixes
@@ -124,7 +145,7 @@ commandProps = T.askOption $ \(MkMaxRuns limit) ->
       H.property $ do
         log@MkLog {cmd = Just (MkCommand _ cmd')} <- H.forAll LGens.genLogWithCmd
         let cmdTxt = "[" <> cmd' <> "]"
-            MkLogText result = Queue.formatFileLog @MockTime log ^. #runMockTime
+            MkLogText result = Queue.formatFileLog @_ @MockTime log ^. #runMockTime
         H.annotate $ T.unpack result
         H.assert $ T.isInfixOf cmdTxt result
 
@@ -134,7 +155,7 @@ shapeProps = T.askOption $ \(MkMaxRuns limit) ->
     H.withTests limit $
       H.property $ do
         log@MkLog {cmd, msg} <- H.forAll LGens.genLogWithCmd
-        let MkLogText result = Queue.formatFileLog @MockTime log ^. #runMockTime
+        let MkLogText result = Queue.formatFileLog @_ @MockTime log ^. #runMockTime
             expected =
               "["
                 <> sysTime